feat(web): rebuild the Hosting tab on the settings kit
Hosting was six bordered boxes with their own header bar each, tables
inside cards inside sections, icon buttons with no tooltip, and details
that snapped open. It now reads like the settings pages it sits next to:
a page title with the live host beside it, one section per concern with
a one-line description, tables where the data is tabular (deployments,
history) and hairline-row cards where it is a list (env, secrets,
domains, redirects).

Two primitives land in @decocms/ui so the next pages start from here:

- `IconButton` — an icon-only button whose `label` is mandatory and
  becomes both the tooltip and the accessible name. Every icon-only
  control on the page is one.
- `CollapsibleContent` now animates height open and closed, the way
  `AccordionContent` already did (tw-animate-css ships the keyframes;
  Radix publishes the measured height). `motion-reduce` turns it back
  into a plain toggle.

Details unfold with that motion everywhere: a deployment's build message
under its table row, a domain's or host redirect's registrar records
under its card row, and the read-only variables declared in code under
the env list. Deploy history shows eight events and offers the rest.

The 2400-line file is a folder now — `api.ts` for wire types and
formatters, `shell.tsx` for the shared frame, one file per section — and
the copy each section needed is in both languages. Nothing about what
the tab fetches or writes changed.

Also: `fmtDuration` rounded seconds after splitting minutes, so 119.6s
printed "1m 60s"; and the dev mock marked every non-failed deployment
`up`, which painted five "Live" badges.
