+++ packages/ui/src/components/collapsible.tsx
@@ -2,6 +2,8 @@
 
 import * as CollapsiblePrimitive from "@radix-ui/react-collapsible";
 
+import { cn } from "../lib/utils.ts";
+
 function Collapsible({
   ...props
 }: React.ComponentProps<typeof CollapsiblePrimitive.Root>) {
@@ -19,12 +21,23 @@ function CollapsibleTrigger({
   );
 }
 
+/**
+ * Opens and closes by animating height, like `AccordionContent`, so a
+ * disclosure never snaps. Radix keeps the node mounted through the exit
+ * animation and publishes the measured height as a CSS variable the keyframes
+ * read; `motion-reduce` turns it back into a plain toggle.
+ */
 function CollapsibleContent({
+  className,
   ...props
 }: React.ComponentProps<typeof CollapsiblePrimitive.CollapsibleContent>) {
   return (
     <CollapsiblePrimitive.CollapsibleContent
       data-slot="collapsible-content"
+      className={cn(
+        "overflow-hidden data-[state=closed]:animate-collapsible-up data-[state=open]:animate-collapsible-down motion-reduce:animate-none",
+        className,
+      )}
       {...props}
     />
   );
